Dev anil pokhariyal

+++ main.py
@@ -1,4 +1,4 @@
-from flask import Flask, request, session, render_template, abort, redirect
+from flask import Flask, request, session, render_template, abort, redirect, jsonify
 from flask_mysqldb import MySQL, MySQLdb
 import os, logging
 
@@ -47,23 +47,26 @@ def create_database():
             return redirect('/')
 
         db_name = request.form.get('database_name')
-        collation = eval(request.form.get('database_collection'))
-        create_db = "CREATE DATABASE " + db_name
-        if collation:
-            create_db += " CHARACTER SET " + collation[1] + " COLLATE " + collation[0] + ";"
-        query(connection, create_db)
-
-        # alter database and it's tables
         alter_db = request.form.get('alter_db_name', None)
-        if alter_db:
-            query(connection, "use information_schema;")
-            table_query = "SELECT TABLE_NAME FROM TABLES WHERE TABLE_SCHEMA = '" + alter_db + "'"
-            tables = query(connection, table_query)
+        if db_name != alter_db:
+            collation = eval(request.form.get('database_collection'))
+            create_db = "CREATE DATABASE " + db_name
+            if collation:
+                create_db += " CHARACTER SET " + collation[1] + " COLLATE " + collation[0] + ";"
+            query(connection, create_db)
 
-            query(connection, "use "+alter_db+";")
-            for table in tables:
-                query(connection, "RENAME TABLE " + table['TABLE_NAME'] + " TO `"+db_name+"`.`"+table['TABLE_NAME']+"`;")
-            query(connection, "DROP DATABASE " + alter_db + ";")
+            # alter database and it's tables
+            if alter_db and alter_db != 'None':
+                query(connection, "use information_schema;")
+                table_query = "SELECT TABLE_NAME FROM TABLES WHERE TABLE_SCHEMA = '" + alter_db + "'"
+                tables = query(connection, table_query)
+
+                query(connection, "use " + alter_db + ";")
+                for table in tables:
+                    query(connection,
+                          "RENAME TABLE " + table['TABLE_NAME'] + " TO `" + db_name + "`.`" + table[
+                              'TABLE_NAME'] + "`;")
+                query(connection, "DROP DATABASE " + alter_db + ";")
 
         return redirect('/')
 
@@ -82,6 +85,29 @@ def drop_database():
             query(connection, "DROP DATABASE " + database + ";")
 
         return redirect('/')
+
+
+@app.route("/create_table", methods=["POST"])
+def create_table():
+    """create table in selected database"""
+    database = request.form.get("database")
+    table_name = request.form.get('table_name')
+    # for mysql engine connection
+    if 'system' in session and session['system'] == 'mysql':
+        try:
+            connection = mysql_connection()
+        except ConnectionError as ex:
+            return redirect('/')
+
+        # use databases and create table
+        try:
+            query(connection, "use " + database + ";")
+            create_query = request.form.get('create_table_query')
+            query(connection, create_query)
+        except (MySQLdb.Error, MySQLdb.Warning) as e:
+            print(e)
+
+    return redirect('/py_adminer?database=' + database + '&table=' + table_name)
 
 
 @app.route("/py_adminer", methods=["GET", "POST"])
@@ -104,6 +130,11 @@ def py_admin():
     selected_table = None
     action = request.args.get('action', None)
     create = request.args.get('create', None)
+    sql_panel = request.args.get('sql_panel', None)
+    if not sql_panel:
+        sql_panel = request.form.get('sql_panel', None)
+    sql_query = request.form.get('sql_query', None)
+    query_output = []
     if 'pass' in session:
         login = True
 
@@ -142,7 +173,7 @@ def py_admin():
             selected_db = database
 
         # options of collation and engine for creating database or table
-        if create:
+        if create or action=='alter':
             # fetching all collations for database collection type on UI
             query(connection, "use information_schema;")
             all_collations = query(connection, "SELECT * FROM COLLATIONS ORDER BY SORTLEN ASC")
@@ -152,18 +183,12 @@ def py_admin():
                 else:
                     db_collations[collation['CHARACTER_SET_NAME']] = [collation, ]
 
-            if create == 'table':
-                # for create table
-                query(connection, "use information_schema;")
-                db_engines = query(connection, "SELECT * FROM ENGINES")
-                column_types = query(connection, "SELECT * FROM COLLATIONS ORDER BY SORTLEN ASC")
-                for collation in all_collations:
-                    if collation['CHARACTER_SET_NAME'] in db_collations:
-                        db_collations[collation['CHARACTER_SET_NAME']].append(collation)
-                    else:
-                        db_collations[collation['CHARACTER_SET_NAME']] = [collation, ]
+            # for create table
+            query(connection, "use information_schema;")
+            db_engines = query(connection, "SELECT * FROM ENGINES")
+            column_types = query(connection, "SELECT * FROM COLLATIONS ORDER BY SORTLEN ASC")
 
-        # fetching tables from selected database
+        # fetching tables from selected database if not create request or sql_panel
         if request.args.get('database') and not create:
             database = str(request.args.get('database'))
             selected_db = database
@@ -174,7 +199,7 @@ def py_admin():
             tables = query(connection, table_query)
 
         # fetching table data and structure
-        if selected_db and request.args.get('table') and not action:
+        if selected_db and request.args.get('table') and (not action or action == 'alter'):
             table_name = str(request.args.get('table'))
             selected_table = table_name
             query(connection, "use information_schema;")
@@ -184,7 +209,7 @@ def py_admin():
         # todo index and foreign keys
 
         # fetching table data
-        if selected_db and request.args.get('table') and action:
+        if selected_db and request.args.get('table') and action == 'data':
             table_name = str(request.args.get('table'))
             selected_table = table_name
             limit = 1000
@@ -197,6 +222,11 @@ def py_admin():
             data_query = "SELECT * FROM " + selected_table + " LIMIT " + str(limit)
             table_data = query(connection, data_query)
 
+        # for sql raw query execute
+        if sql_query:
+            query(connection, "use " + selected_db)
+            query_output = query(connection, sql_query)
+
         # required in case query fails
         if databases:
             session['pass'] = True
@@ -206,7 +236,25 @@ def py_admin():
                            mysql_version=mysql_version, create=create, action=action,
                            tables=tables, table_structure=table_structure, db_engines=db_engines,
                            table_data=table_data, table_columns=table_columns, db_collations=db_collations,
-                           selected_db=selected_db, selected_table=selected_table)
+                           selected_db=selected_db, selected_table=selected_table, sql_panel=sql_panel,
+                           sql_query=sql_query, query_output=query_output)
+
+
+@app.route("/logout", methods=["GET", "POST"])
+def logout():
+    if 'system' in session:
+        session.pop('system')
+    if 'host' in session:
+        session.pop('host')
+    if 'user' in session:
+        session.pop('user')
+    if 'password' in session:
+        session.pop('password')
+    if 'database' in session:
+        session.pop('database')
+    if 'pass' in session:
+        session.pop('pass')
+    return redirect('/')
 
 
 if __name__ == '__main__':
